Centralize grid line overlay sync and word-connector SVG restyle
Pull shared overlay sizing (optional beforeMeasure), swipe pixel lock,
and gradient line restyling into js/grid-layout.js and
js/word-connector-lines.js; use deferTwoAnimationFrames in shift-dom.

Trim redundant JSDoc on the helpers.
